1_Paciente/django_sample/sample/requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
django-filter==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
drf-yasg==1.4.3
flex==6.12.0
future==0.16.0
gunicorn==19.7.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==2.0
jsonschema==2.6.0
kombu==4.1.0
MarkupSafe==1.0
numpy==1.14.1
openapi-codec==1.3.2
psycopg2==2.7.4
python-env==1.0.0
pytz==2018.3
PyYAML==3.12
raven==6.5.0
redis==2.10.6
requests==2.18.4
rfc3987==1.3.7
ruamel.yaml==0.15.35
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
vine==1.1.4
django-cors-headers
mysqlclient